updated shopper profile

+++ core/templatetags/tag_library.py
@@ -4,10 +4,13 @@ register = template.Library()
 
 @register.filter()
 def to_int(value):
-    return float(value)
+
+    return int(value)
+
 
 @register.filter()
 def to_float(value):
+
     return float(value)
 
 @register.filter()
